fix 'org.etudes.tool.melete.MeleteSiteAndUserInfo.processNavigate java.lang.ClassCastException: java.lang.Long cannot be cast to java.lang.Integer' query.uniqueResult(); was returning Long instead of Int. Maybe due to newer version of Hibernate?

+++ melete-api/src/java/org/etudes/api/app/melete/ModuleService.java
@@ -191,7 +191,7 @@ public interface ModuleService
 	 *        The course Id
 	 * @return
 	 */
-	public int getCourseModuleSize(String courseId);
+	public long getCourseModuleSize(String courseId);
 
 	/**
 	 * Gets the earliest start date or end date(if defined) of all the modules. Returns null if dates don't exist
